requirements.txt
# Core
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.1
seaborn==0.13.2
requests==2.32.3
python-dotenv==1.0.1

# Database
SQLAlchemy==2.0.32
mysql-connector-python==9.0.0

# Machine Learning
scikit-learn==1.5.1
joblib==1.5.1
xgboost==2.1.0
shap==0.45.1

# API
fastapi==0.111.1
pydantic==2.11.7
uvicorn==0.30.3

# Scheduling (optional for daily jobs)
APScheduler==3.10.4
